+++ core/src/message.rs
@@ -18,7 +18,7 @@
 //! See the README.md for more information on how to document messages.
 extern crate alloc;
 
-use crate::{AccountId, Balance, Bytes128, CheckpointId, Id, ProjectName, H256};
+use crate::{AccountId, Balance, Bytes128, CheckpointId, Id, ProjectDomain, ProjectName, H256};
 use alloc::prelude::v1::Vec;
 use parity_scale_codec::{Decode, Encode};
 
@@ -128,23 +128,24 @@ pub struct RegisterMember {
 ///
 /// # State-dependent validations
 ///
-/// The involved org must exit.
+/// The involved project domain must exit.
 ///
 /// A user associated with the author must exist.
 ///
-/// The user associated with the author must be a member of the involved org.
+/// The user associated with the author must a member of
+/// the involved org, when an org is specified as the project domain.
 ///
 /// A checkpoint with the given ID must exist.
 ///
-/// A project with the same name must not yet exist in the org.
+/// A project with the same name must not yet exist in domain.
 ///
 #[derive(Decode, Encode, Clone, Debug, Eq, PartialEq)]
 pub struct RegisterProject {
-    // The name of the project to register, unique in the org.
+    // The name of the project to register, unique under its domain.
     pub project_name: ProjectName,
 
-    /// The org in which to register the project.
-    pub org_id: Id,
+    /// The domain of the project.
+    pub project_domain: ProjectDomain,
 
     /// Initial checkpoint of the project.
     pub checkpoint_id: CheckpointId,
@@ -186,7 +187,7 @@ pub struct CreateCheckpoint {
 #[derive(Decode, Encode, Clone, Debug, Eq, PartialEq)]
 pub struct SetCheckpoint {
     pub project_name: ProjectName,
-    pub org_id: Id,
+    pub project_domain: ProjectDomain,
     pub new_checkpoint_id: CheckpointId,
 }