+++ runtime-tests/tests/project_registration.rs
@@ -22,7 +22,7 @@ use radicle_registry_client::*;
 use radicle_registry_test_utils::*;
 
 #[async_std::test]
-async fn register_project() {
+async fn register_project_under_org() {
     let (client, _) = Client::new_emulator();
     let (author, _) = key_pair_with_associated_user(&client).await;
 
@@ -51,24 +51,28 @@ async fn register_project() {
     let initial_balance = 1000;
     transfer(&client, &author, org.account_id, initial_balance).await;
 
+    let project_domain = ProjectDomain::Org(register_org.org_id.clone());
     let random_fee = random_balance();
-    let message = random_register_project_message(register_org.org_id.clone(), checkpoint_id);
+    let message = random_register_project_message(&project_domain, checkpoint_id);
     let tx_included = submit_ok_with_fee(&client, &author, message.clone(), random_fee).await;
 
     let project = client
-        .get_project(message.clone().project_name, message.clone().org_id)
+        .get_project(message.project_name.clone(), message.project_domain.clone())
         .await
         .unwrap()
         .unwrap();
     assert_eq!(project.name.clone(), message.project_name.clone());
-    assert_eq!(project.org_id.clone(), message.org_id.clone());
+    assert_eq!(project.domain.clone(), message.project_domain.clone());
     assert_eq!(project.current_cp.clone(), checkpoint_id);
     assert_eq!(project.metadata.clone(), message.metadata.clone());
 
     assert_eq!(
         tx_included.events[0],
-        RegistryEvent::ProjectRegistered(message.clone().project_name, message.clone().org_id)
-            .into()
+        RegistryEvent::ProjectRegistered(
+            message.project_name.clone(),
+            message.project_domain.clone()
+        )
+        .into()
     );
 
     let has_project = client
@@ -76,7 +80,7 @@ async fn register_project() {
         .await
         .unwrap()
         .iter()
-        .any(|id| *id == (message.project_name.clone(), message.org_id.clone()));
+        .any(|id| *id == (message.project_name.clone(), message.project_domain.clone()));
     assert!(has_project, "Registered project not found in project list");
 
     let checkpoint_ = state::Checkpoint {
@@ -105,7 +109,7 @@ async fn register_project() {
 }
 
 #[async_std::test]
-async fn register_project_with_inexistent_org() {
+async fn register_project_under_inexistent_org() {
     let (client, _) = Client::new_emulator();
     let (author, _) = key_pair_with_associated_user(&client).await;
 
@@ -123,7 +127,8 @@ async fn register_project_with_inexistent_org() {
     .unwrap();
 
     let inexistent_org_id = random_id();
-    let message = random_register_project_message(inexistent_org_id, checkpoint_id);
+    let message =
+        random_register_project_message(&ProjectDomain::Org(inexistent_org_id), checkpoint_id);
     let tx_included = submit_ok(&client, &author, message.clone()).await;
 
     assert_eq!(tx_included.result, Err(RegistryError::InexistentOrg.into()));
@@ -156,7 +161,8 @@ async fn register_project_with_duplicate_id() {
     let org = client.get_org(org_id.clone()).await.unwrap().unwrap();
     transfer(&client, &author, org.account_id, 1000).await;
 
-    let message = random_register_project_message(org_id.clone(), checkpoint_id);
+    let message =
+        random_register_project_message(&ProjectDomain::Org(org_id.clone()), checkpoint_id);
     submit_ok(&client, &author, message.clone()).await;
 
     // Duplicate submission with a different metadata.
@@ -176,7 +182,7 @@ async fn register_project_with_duplicate_id() {
     );
 
     let project = client
-        .get_project(message.project_name, message.org_id)
+        .get_project(message.project_name, message.project_domain)
         .await
         .unwrap()
         .unwrap();
@@ -201,7 +207,8 @@ async fn register_project_with_bad_checkpoint() {
     let checkpoint_id = H256::random();
 
     let org_id = random_id();
-    let register_project = random_register_project_message(org_id.clone(), checkpoint_id);
+    let project_domain = ProjectDomain::Org(org_id.clone());
+    let register_project = random_register_project_message(&project_domain, checkpoint_id);
     let register_org = message::RegisterOrg {
         org_id: org_id.clone(),
     };
@@ -219,7 +226,7 @@ async fn register_project_with_bad_checkpoint() {
     );
 
     assert!(client
-        .get_project(register_project.project_name, register_project.org_id)
+        .get_project(register_project.project_name, project_domain)
         .await
         .unwrap()
         .is_none());
@@ -240,7 +247,8 @@ async fn register_project_with_bad_actor() {
 
     // The bad actor attempts to register a project within that org.
     let initial_balance = client.free_balance(&bad_actor.public()).await.unwrap();
-    let register_project = random_register_project_message(org_id.clone(), H256::random());
+    let project_domain = ProjectDomain::Org(org_id.clone());
+    let register_project = random_register_project_message(&project_domain, H256::random());
     let random_fee = random_balance();
     let tx_included =
         submit_ok_with_fee(&client, &bad_actor, register_project.clone(), random_fee).await;
@@ -258,7 +266,7 @@ async fn register_project_with_bad_actor() {
     );
 
     assert!(client
-        .get_project(register_project.project_name, register_project.org_id)
+        .get_project(register_project.project_name, project_domain)
         .await
         .unwrap()
         .is_none());